requirements-web.txt
# PlantLLM Webアプリケーション用の依存関係
# Usage: pip install -r requirements-web.txt
# Updated: 2026-01-18 (context7で最新バージョン確認済み)

# Web Framework
fastapi>=0.128.0
uvicorn>=0.34.0
starlette>=0.45.0
python-multipart>=0.0.20

# Database
SQLAlchemy>=2.0.45
aiosqlite>=0.20.0
greenlet>=3.1.1

# Pydantic
pydantic>=2.12.5
pydantic_core>=2.27.2
annotated-types>=0.7.0

# Google GenAI SDK (新SDK - google-generativeaiは非推奨)
google-genai>=1.0.0
google-api-core>=2.24.0
google-auth>=2.47.0
googleapis-common-protos>=1.66.0
grpcio>=1.69.0
grpcio-status>=1.62.3
proto-plus>=1.25.0
protobuf>=5.29.0

# Image Processing
pillow>=12.1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=9.0.0
aiofiles>=24.1.0

# Network
requests>=2.32.3
certifi>=2024.12.14
charset-normalizer>=3.4.1
idna>=3.10
urllib3>=2.3.0

# Async
anyio>=4.8.0
sniffio>=1.3.1
h11>=0.16.0

# Other
click>=8.1.8
typing_extensions>=4.12.2
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
rsa>=4.9
cachetools>=5.5.0

# Testing (dev)
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
